builds/hawk-centos7.spec: Fixes for the install and upgrade processes
During upgrade of hawk, we are deleting the hawk user, which deletes
the sqlite file, that was initially created during installation.
It also leaves the hawk group, which results in issues during upgrade
and install.

In addition we were missing the creation of the cache dir, when we
were upgrading, because %postun was called without check if it was
un upgrade or uninstall.

The fixes here try to solve the above provlems.

+++ builds/hawk-centos7.spec
@@ -86,8 +86,11 @@ fi
 
 # Initialize the Hawk SQLite DB
 if [ ! -f /var/cache/hawk/hawk.sqlite ]; then
+	if [ ! -d /var/cache/hawk ]; then
+		mkdir -p /var/cache/hawk
+	fi
 	sqlite3 /var/cache/hawk/hawk.sqlite < /usr/share/hawk/hawk_db.sqlite
-	chown hawk: /usr/share/hawk/hawk_db.sqlite
+	chown hawk: /var/cache/hawk/hawk.sqlite
 fi
 
 %preun
@@ -98,8 +101,11 @@ fi
 %postun
 %systemd_postun_with_restart hawk.service
 
-if getent passwd hawk > /dev/null; then
-	userdel -r hawk
+if [ "$1" == "0" ]; then
+	if getent passwd hawk > /dev/null; then
+		userdel -r hawk
+		groupdel hawk
+	fi
 fi
 
 %posttrans
